pyproject.toml
[project]
name = "my-langchain-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.11.0",
    "isort>=6.1.0",
    "langchain>=0.3.28",
    "langchain-anthropic>=0.3.22",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.84",
    "langchain-ollama>=0.3.10",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.23",
]
